requirements.txt
streamlit
python-docx
python-pptx
openai
requests
Pillow
torch
transformers
ffmpeg
torchvision
scikit-learn
pandas
numpy
matplotlib
seaborn
plotly
streamlit-ace
streamlit-option-menu
streamlit-ace
streamlit-embedcode
torch
transformers